+++ packages/xd-crossword-tools/src/getWordTilesForCursor.ts
@@ -1,4 +1,4 @@
-import { CrosswordJSON, Cursor, Position, getTile } from "xd-crossword-parser"
+import { CrosswordJSON, Cursor, Position, getTile } from "xd-crossword-tools-parser"
 
 /** Gets all the related tiles for the cursor, may contain dupes */
 export const getWordTilesForCursor = (tiles: CrosswordJSON["tiles"], cursor: Cursor) => {