+++ CHANGELOG.md
@@ -2,7 +2,7 @@ This isn't a comprehensive doc because to our knowledge there are no OSS consume
 
 ### 9.0.0
 
-- Split out the parser into it's own package. So, if you're making a crossword game, then you can just depend on `xd-crossword-parser` and not have to depend on the tools.
+- Split out the parser into it's own package. So, if you're making a crossword game, then you can just depend on `xd-crossword-tools-parser` and not have to depend on the tools.
 
 ### 8.1.0